[chore] Allow dead code
Most of the dead code here is related to allowing the use of uris rather
than files.  I really don't want this code being a pest and hammering
the various sites that it knows how to parse, so I don't want to enable
uris until there's some sort of caching mechanism.