[Feat] includes all the black Friday results.  Closes #19.
I just saved all of the Black Friday Turkey Trots results and fixed
the web_scorer parser to handle 2019, where instead of an overall
table there is a female table and a male table.  Combining new results
and a parser fix in a single commit is slightly impolite, but I'm
doing this in the middle of the night because I drank too much
caffeine earlier in the day.